// src/app/core/services/models/registration-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegistrationRequest {
  dateOfBirth?: string;
  email: string;
  firstname: string;
  gender?: 'MALE' | 'FEMALE';
  lastname: string;
  password: string;
  skill?: 'COMMUNICATION' | 'ORGANIZATION' | 'IT' | 'TEACHING' | 'FIRST_AID' | 'LOGISTICS' | 'PSYCHOLOGICAL_SUPPORT' | 'LEADERSHIP';
}